Refactor the user profile route (DESC)
Moved the user profile route into a separate route structure. /submission/u/[userId] -> user/[userId]
