+++ js/src/app/user/[userId]/_components/UserSubmissionContent.tsx
@@ -1,5 +1,5 @@
-import SearchBox from "@/app/submission/u/[userId]/_components/SearchBox";
-import { useUserSubmissionsQuery } from "@/app/submission/u/[userId]/hooks";
+import SearchBox from "@/app/user/[userId]/_components/SearchBox";
+import { useUserSubmissionsQuery } from "@/app/user/[userId]/hooks";
 import CustomPagination from "@/components/ui/CustomPagination";
 import { Footer } from "@/components/ui/footer/Footer";
 import Header from "@/components/ui/header/Header";
